Update README.md
# This project is part of the Machine Learning and Deep Learning ( MLDL 2021 ) course, Master of Data Science and Engineering, Politecnico Di Torino 
# Abstract 
In deep CNN-based models for semantic segmentation, high accuracy relies on rich spatial context (large receptive fields) and fine spatial details (high resolution), both
of which incur high computational costs. The other limitation is the fact that these models are based on supervision with pixel-level ground truth but may not generalize well to unseen image domains. Our goal is to create a model that generalizes well and adapt source ground truth labels to target domain labels while maintaining lower computations and inference time. In our paper, we combined a real-time semantic segmentation model with an unsupervised adversarial domain adaptation approach and we used a domain classifier to align the features extracted from the output space. As the labeling process is tedious and labor-intensive, we adopted a synthetic-to-real scenario. To meet our objective, we researched the effect of using a lightweight discriminator network to enhance the time and avoid over-fitting by reducing the capacity of the model. We compared three different discriminator architectures and we accomplished significant enhancement both in terms of speed and segmentation performance.